// frontend/src/client/models/UserOut.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UserOut = {
    email: string;
    is_active?: boolean;
    is_superuser?: boolean;
    is_technical_staff?: (boolean | null);
    full_name?: (string | null);
    phone?: (string | null);
    room?: (string | null);
    study_group?: (string | null);
    institute?: (string | null);
    id: number;
};