// frontend/src/client/models/UserUpdate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UserUpdate = {
    email?: (string | null);
    is_active?: boolean;
    is_superuser?: boolean;
    full_name?: (string | null);
    phone?: (string | null);
    room?: (string | null);
    study_group?: (string | null);
    institute?: (string | null);
    password?: (string | null);
};